+++ src/migrations/Install.php
@@ -1,0 +1,24 @@
+<?php
+namespace burnthebook\ClickCollect\migrations;
+
+use Craft;
+use craft\db\Migration;
+use craft\db\Query;
+use craft\db\Table;
+use craft\elements\Entry;
+use craft\helpers\ArrayHelper;
+use craft\helpers\Json;
+use craft\helpers\MigrationHelper;
+use craft\helpers\StringHelper;
+use craft\services\Plugins;
+
+class Install extends Migration
+{
+    // Public Methods
+    // =========================================================================
+
+    public function safeUp()
+    {
+
+    }
+}