+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-# 0.0.1
+# Release notes for Click & Collect for Craft Commerce
 
-Plugin init
+# Click & Collect for Craft Commerce 0.0.1 - 2022-07-01
+
+> {note} This plugin is not yet ready for public release or consumption
+
+Initial Alpha and Dev of Click & Collect for Craft Commerce Plugin
+
+- Adds Settings Page
+- Adds "Click & Collect" Shipping Method
+- 
+